python/projectNew/requirements.txt
Django==2.0.3
PyMySQL==0.8.0
djangorestframework==3.8.2
